Did final file refactor before TS refactor
