+++ src/App.js
@@ -3,17 +3,15 @@ import {
     BrowserRouter as Router,
     Switch,
     Route,
-    Link
 } from "react-router-dom";
 
-import metarParse from "aewx-metar-parser";
-
-import Toolbar from "./Toolbar.js";
+import Toolbar from "./toolbar/Toolbar.js";
 import PreparationWidgets from "./preparationWidget/PreparationWidget.js";
 import Settings from "./settings/Settings.js";
 import Profile from "./profile/Profile.js";
 
 import './App.scss';
+import './toolbar/Toolbar.scss';
 import './settings/Settings.scss';
 import './profile/Profile.scss';
 import './preparationWidget/PreparationWidget.scss';
