+++ src/App.scss
@@ -54,79 +54,3 @@ body {
   margin-bottom: 0.75%;
   color: var(--mainText)
 }
-
-.Toolbar {
-  $padding: 1em;
-  $spacing-padding: 2.25rem;
-
-  background: var(--cardBackground);
-
-  width: 100vw;
-
-  position: fixed;
-  top: 0;
-  padding: $padding $padding $padding $spacing-padding;
-  margin-right: 5em;
-  display: flex;
-  flex-direction: row;
-  justify-content: flex-start;
-  align-items: center;
-
-  #Tabs {
-    display: flex;
-    flex-direction: row;
-    justify-content: flex-start;
-    align-items: flex-start;
-
-    margin-left: $spacing-padding;
-
-    img.FBW-Logo {
-      width: 2.25em;
-    }
-
-    .Tab {
-      margin-top: .4em;
-      margin-right: $spacing-padding;
-      padding-bottom: .135em;
-
-      font-family: "Nunito Sans", sans-serif;
-      font-size: 1.5rem;
-      font-weight: 700;
-
-      color: var(--mainText);
-
-      cursor: pointer;
-
-      &.Active {
-        border-bottom: solid var(--mainTextSecondary) .1em;
-      }
-
-      a {
-        color: inherit;
-        text-decoration: inherit;
-      }
-    }
-  }
-
-  #ProfileToolbar {
-    margin-left: auto;
-    display: flex;
-    align-items: left;
-    font-weight: 600;
-    font-size: 1.3rem;
-    color: var(--profileText);
-
-    #Icon {
-      margin-left: 10px;
-      font-size: 3rem;
-    }
-
-    &.Active {
-      border-bottom: solid var(--mainTextSecondary) .1em;
-    }
-  }
-
-  #Icon {
-    justify-content: end;
-  }
-}
